Fix deprecated methods
